+++ trail-register/src/main/java/com/rei/trailregister/UsageRepository.java
@@ -24,7 +24,7 @@ public interface UsageRepository {
     List<String> getEnvironments(String app);
     List<String> getCategories(String app, String env);
     List<String> getKeys(String app, String env, String category);
-    Map<String, Integer> getUsagesByDate(UsageKey key, int days);
+    Map<String, Long> getUsagesByDate(UsageKey key, int days);
     long getUsages(UsageKey key, int days);
     
     default void runCompaction() {}
