+++ trail-register/src/test/java/com/rei/trailregister/FileUsageRepositoryTest.java
@@ -55,7 +55,7 @@ public class FileUsageRepositoryTest {
 		repo.recordUsages(usageKey, 2, LocalDate.now().minusDays(2));
 		assertEquals(379, repo.getUsages(usageKey, 366));
 		
-		Map<String, Integer> usagesByDate = repo.getUsagesByDate(usageKey, 366);
+		Map<String, Long> usagesByDate = repo.getUsagesByDate(usageKey, 366);
         System.out.println(usagesByDate);
         assertEquals(366, usagesByDate.size());
